🔨feat-user: user->auth로 로그인 기능 파일구조 변경
