+++ src/auth/decorator/get-user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { User } from '../model/user.entity';
+import { User } from '../../user/model/user.entity';
 import { plainToInstance } from 'class-transformer';
 
 export const GetUser = createParamDecorator(
