+++ src/user/controller/user.controller.ts
@@ -2,7 +2,6 @@ import {
   Body,
   Controller,
   Post,
-  Req,
   UseGuards,
   ValidationPipe,
 } from '@nestjs/common';
@@ -10,9 +9,7 @@ import { UserService } from '../service/user.service';
 import { User } from '../model/user.entity';
 import { ApiTags } from '@nestjs/swagger';
 import { CreateUserDto } from '../dto/create-user.dto';
-import { AuthCredentialsDto } from '../dto/authCredentials.dto';
 import { AuthGuard } from '@nestjs/passport';
-import { GetUser } from '../decorator/get-user.decorator';
 
 @ApiTags('User')
 @Controller('users')
@@ -24,16 +21,9 @@ export class UserController {
     return this.userService.signup(createUserDto);
   }
 
-  @Post('/login')
-  signin(
-    @Body(ValidationPipe) authCredentialsDto: AuthCredentialsDto,
-  ): Promise<{ accessToken: string }> {
-    return this.userService.signin(authCredentialsDto);
-  }
-
-  @UseGuards(AuthGuard())
-  @Post('/test')
-  test(@GetUser() user: User) {
-    console.log(user);
-  }
+  // @UseGuards(AuthGuard())
+  // @Post('/test')
+  // test(@GetUser() user: User) {
+  //   console.log(user);
+  // }
 }
