+++ src/user/service/user.service.ts
@@ -8,15 +8,10 @@ import { CreateUserDto } from '../dto/create-user.dto';
 import { User } from '../model/user.entity';
 import * as brcypt from 'bcryptjs';
 import { plainToInstance } from 'class-transformer';
-import { AuthCredentialsDto } from '../dto/authCredentials.dto';
-import { JwtService } from '@nestjs/jwt';
 
 @Injectable()
 export class UserService {
-  constructor(
-    private readonly userRepository: UserRepository,
-    private readonly jwtService: JwtService,
-  ) {}
+  constructor(private readonly userRepository: UserRepository) {}
 
   // 회원가입
   async signup(createUserDto: CreateUserDto): Promise<User> {
@@ -44,27 +39,6 @@ export class UserService {
       return plainToInstance(User, user, {
         excludeExtraneousValues: false,
       });
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  // 로그인
-  async signin(
-    authCredentialsDto: AuthCredentialsDto,
-  ): Promise<{ accessToken: string }> {
-    const { userId, userPassword } = authCredentialsDto;
-    const user = await this.userRepository.findByUserId(userId);
-    const isCorrect = await brcypt.compare(userPassword, user.userPassword);
-
-    if (!user) throw new UnauthorizedException('존재하지 않는 아이디입니다.');
-    if (!isCorrect) throw new UnauthorizedException('비밀번호가 틀렸습니다.');
-
-    try {
-      const payload = { userId };
-      const accessToken = this.jwtService.sign(payload);
-
-      return { accessToken };
     } catch (error) {
       console.log(error);
     }
